fix(codegen): Use i32 for nested function Type::Int to match rest of codebase (Refs DEPYLER-0759)
Root Cause: hir_type_to_tokens() in stmt_gen.rs used i64 for Type::Int while all
other type mappers (codegen.rs, expr_gen.rs, argparse_transform.rs, etc.) use i32.
This caused E0308 type mismatch errors when nested functions captured i32 variables
from outer scope but declared their parameters as i64.

Fix: Changed Type::Int => quote! { i64 } to Type::Int => quote! { i32 } in
hir_type_to_tokens() to maintain type consistency across all code generation paths.

+++ crates/depyler-core/src/rust_gen/stmt_gen.rs
@@ -6503,7 +6503,11 @@ pub(crate) fn hir_type_to_tokens(ty: &Type, _ctx: &CodeGenContext) -> proc_macro
     use quote::quote;
 
     match ty {
-        Type::Int => quote! { i64 },
+        // DEPYLER-0759: Use i32 to match all other type mappers in the codebase
+        // (codegen.rs, expr_gen.rs, argparse_transform.rs, etc.)
+        // Using i64 here caused E0308 mismatched types when nested functions
+        // captured i32 variables from outer scope
+        Type::Int => quote! { i32 },
         Type::Float => quote! { f64 },
         Type::String => quote! { String },
         Type::Bool => quote! { bool },
